+++ src/components/AdminPane/Manage/ManageChallenges/EditChallenge/Schemas/PrioritiesSchema.js
@@ -1,32 +1,30 @@
 import { TaskPriority,
          taskPriorityLabels }
-       from '../../../../../services/Task/TaskPriority/TaskPriority'
+       from '../../../../../../services/Task/TaskPriority/TaskPriority'
 import _map from 'lodash/map'
 import _values from 'lodash/values'
-import messages from './Messages'
+import messages from '../Messages'
+
+const STEP_ID = "Priorities"
 
 /**
- * Generates a JSON Schema describing Step 3 (task priorities) of Edit
- * Challenge workflow intended for consumption by react-jsonschema-form.
+ * Generates a JSON Schema describing priority fields of Edit
+ * Challenge workflow intended for consumption by react-jsonschema-form
  *
  * > Note that react-jsonschema-form only presents values for checkbox fields
  * > if they are checked, so it's best to specify radio buttons in the uiSchema
- * > for boolean fields if additional post-processing is to be avoided.
- *
- * @param intl - intl instance from react-intl
+ * > for boolean fields if additional post-processing is to be avoided
  *
  * @see See http://json-schema.org
  * @see See https://github.com/mozilla-services/react-jsonschema-form
  *
  * @author [Neil Rotstan](https://github.com/nrotstan)
  */
-export const jsSchema = intl => {
+export const jsSchema = (intl, user, challengeData, extraErrors, options={}) => {
   const localizedPriorityLabels = taskPriorityLabels(intl)
 
   return {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    title: intl.formatMessage(messages.step3Label),
-    description: intl.formatMessage(messages.step3Description),
     type: "object",
     definitions: {
       tagRule: {
@@ -147,6 +145,7 @@ export const jsSchema = intl => {
         enum: _values(TaskPriority),
         enumNames: _map(TaskPriority, (value, key) => localizedPriorityLabels[key]),
         default: TaskPriority.high,
+        description: intl.formatMessage(messages.defaultPriorityDescription),
       },
       highPriorityRules: {
         title: intl.formatMessage(messages.highPriorityRulesLabel),
@@ -181,7 +180,7 @@ export const jsSchema = intl => {
  *
  * @private
  */
-const priorityRuleGroupUISchema = {
+const priorityRuleGroupUISchema = (isCollapsed) => ({
   classNames: "priority-rule-group",
   condition: {
     "ui:widget": "select",
@@ -192,38 +191,56 @@ const priorityRuleGroupUISchema = {
       classNames: "priority-rule",
       keyType: {
         "ui:widget": "select",
+        "ui:displayLabel": false,
+        "ui:collapsed": isCollapsed,
       },
       valueType: {
         "ui:widget": "select",
+        "ui:displayLabel": false,
+        "ui:collapsed": isCollapsed,
       },
       key: {
         "ui:placeholder": "Property Name",
+        "ui:displayLabel": false,
+        "ui:collapsed": isCollapsed,
       },
       operator: {
         "ui:widget": "select",
+        "ui:displayLabel": false,
+        "ui:collapsed": isCollapsed,
       },
       value: {
         "ui:placeholder": "Property Value",
+        "ui:displayLabel": false,
+        "ui:collapsed": isCollapsed,
       },
       ruleGroup: {
-        classNames: "nested-rule-group mr-border mr-border-green-light mr-p-2 mr-mt-4 mr-flex mr-w-full",
+        classNames: "nested-rule-group mr-border mr-border-white-25 mr-p-2 mr-mt-4 mr-flex mr-w-full",
         rules: {
           items: {
             key: {
               "ui:placeholder": "Property Name",
+              "ui:displayLabel": false,
+              "ui:collapsed": isCollapsed,
             },
             value: {
               "ui:placeholder": "Property Value",
+              "ui:displayLabel": false,
+              "ui:collapsed": isCollapsed,
             },
             ruleGroup: {
-              classNames: "nested-rule-group mr-border mr-border-green-light mr-p-2 mr-mt-4 mr-flex mr-w-full",
+              classNames: "nested-rule-group mr-border mr-border-white-25 mr-p-2 mr-mt-4 mr-flex mr-w-full",
               rules: {
                 items: {
                   key: {
                     "ui:placeholder": "Property Name",
+                    "ui:displayLabel": false,
+                    "ui:collapsed": isCollapsed,
                   },
                   value: {
                     "ui:placeholder": "Property Value",
+                    "ui:displayLabel": false,
+                    "ui:collapsed": isCollapsed,
                   },
                 },
               },
@@ -234,30 +251,41 @@ const priorityRuleGroupUISchema = {
       "ui:order": [ "valueType", "key", "operator", "value", "*" ],
     },
   },
-}
+})
 
 /**
  * uiSchema configuration to assist react-jsonschema-form in determining
- * how to render the schema fields.
+ * how to render the schema fields
  *
  * @see See https://github.com/mozilla-services/react-jsonschema-form
  *
- * > Note: for anything other than text inputs, specifying the ui:widget type in
- * > the form configuration will help the Bulma/RJSFFormFieldAdapter generate the
- * > proper Bulma-compliant markup.
+ * > Note: for anything other than text inputs, specifying the ui:widget type
+ * > in the form configuration will help the RJSFFormFieldAdapter generate the
+ * > proper markup
  */
-export const uiSchema = intl => ({
-  defaultPriority: {
-    "ui:widget": "select",
-    "ui:help": intl.formatMessage(messages.defaultPriorityDescription),
-  },
-  highPriorityRules: {
-    ruleGroup: priorityRuleGroupUISchema,
-  },
-  mediumPriorityRules: {
-    ruleGroup: priorityRuleGroupUISchema,
-  },
-  lowPriorityRules: {
-    ruleGroup: priorityRuleGroupUISchema,
-  },
-})
+export const uiSchema = (intl, user, challengeData, extraErrors, options={}) => {
+  const isCollapsed = options.longForm && (options.collapsedGroups || []).indexOf(STEP_ID) !== -1
+  const toggleCollapsed = options.longForm && options.toggleCollapsed ? () => options.toggleCollapsed(STEP_ID) : undefined
+
+  return {
+    defaultPriority: {
+      "ui:widget": "select",
+      "ui:help": intl.formatMessage(messages.defaultPriorityDescription),
+      "ui:collapsed": isCollapsed,
+      "ui:toggleCollapsed": toggleCollapsed,
+      "ui:groupHeader": options.longForm ? intl.formatMessage(messages.prioritiesStepHeader) : undefined,
+    },
+    highPriorityRules: {
+      ruleGroup: priorityRuleGroupUISchema(isCollapsed),
+      "ui:collapsed": isCollapsed,
+    },
+    mediumPriorityRules: {
+      ruleGroup: priorityRuleGroupUISchema(isCollapsed),
+      "ui:collapsed": isCollapsed,
+    },
+    lowPriorityRules: {
+      ruleGroup: priorityRuleGroupUISchema(isCollapsed),
+      "ui:collapsed": isCollapsed,
+    },
+  }
+}