+++ src/pages/Social/Messages.js
@@ -1,0 +1,46 @@
+import { defineMessages } from 'react-intl'
+
+/**
+ * Internationalized messages for use with Social page
+ */
+export default defineMessages({
+  featuredChallengesHeader: {
+    id: "Social.featuredChallenges.header",
+    defaultMessage: "Featured Challenges",
+  },
+
+  featuredLead: {
+    id: "Social.featuredChallenges.lead",
+    defaultMessage: "Featured on MapRoulette:",
+  },
+
+  newestChallengesHeader: {
+    id: "Social.newestChallenges.header",
+    defaultMessage: "Newest Challenges",
+  },
+
+  newestLead: {
+    id: "Social.newestChallenges.lead",
+    defaultMessage: "New on MapRoulette:",
+  },
+
+  popularChallengesHeader: {
+    id: "Social.popularChallenges.header",
+    defaultMessage: "Popular Challenges",
+  },
+
+  popularLead: {
+    id: "Social.popularChallenges.lead",
+    defaultMessage: "Popular on MapRoulette:",
+  },
+
+  copy: {
+    id: "Social.copy",
+    defaultMessage: "",
+  },
+
+  none: {
+    id: "Social.none",
+    defaultMessage: "None",
+  },
+})